mc_motor: bugfix. fixed speed overflow

+++ lib/mc_motor/mc_motor.cpp
@@ -49,7 +49,7 @@ esp_err_t mc_motor_advanced_init(mc_motor_dev_t *dev, mc_motor_config_t *config)
     return ESP_OK;
 }
 
-esp_err_t mc_motor_set_speed(mc_motor_dev_t *dev, int8_t speed)
+esp_err_t mc_motor_set_speed(mc_motor_dev_t *dev, int16_t speed)
 {
     speed = constrain(speed, -100, 100);
     if (speed == 0)
